requirements.txt
Flask==2.3.3
gunicorn==21.2.0
pypdf==3.17.4
Pillow==10.2.0
PyMuPDF==1.27.1
